requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0

# File upload support
pypdf>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
